#169 Refactor and improve up system
