+++ app/src/main/java/ch/epfl/sweng/zuluzulu/Adapters/PostArrayAdapter.java
@@ -30,11 +30,11 @@ public class PostArrayAdapter extends ArrayAdapter<Post> {
 
     private Post currentPost;
     private TextView timeAgo;
+    private TextView nbUps;
     private ImageView upButton;
     private ImageView downButton;
 
     private FirebaseFirestore db = FirebaseFirestore.getInstance();
-    private DocumentReference documentReference;
 
     public PostArrayAdapter(@NonNull Context context, List<Post> list) {
         super(context, 0, list);
@@ -56,10 +56,8 @@ public class PostArrayAdapter extends ArrayAdapter<Post> {
         timeAgo = view.findViewById(R.id.post_time_ago_textview);
         upButton = view.findViewById(R.id.post_up_button);
         downButton = view.findViewById(R.id.post_down_button);
-        TextView nbUps = view.findViewById(R.id.post_nb_ups_textview);
+        nbUps = view.findViewById(R.id.post_nb_ups_textview);
         TextView nbResponses = view.findViewById(R.id.post_nb_responses_textview);
-
-        documentReference = db.collection("channels/channel" + currentPost.getChannel().getId() + "/posts").document(currentPost.getId());
 
         linearLayout.setBackgroundColor(Color.parseColor(currentPost.getColor()));
         message.setText(currentPost.getMessage());
@@ -72,7 +70,8 @@ public class PostArrayAdapter extends ArrayAdapter<Post> {
         nbUps.setText("" + currentPost.getNbUps());
         nbResponses.setText("" + currentPost.getNbResponses());
 
-        updateUpButtons();
+        setUpUpDownButtons(currentPost, upButton, downButton, nbUps);
+        updateUps(currentPost, upButton, downButton);
 
         return view;
     }
@@ -96,12 +95,54 @@ public class PostArrayAdapter extends ArrayAdapter<Post> {
         }
     }
 
-    private void updateUpButtons() {
-        if (currentPost.isUpByUser()) {
+    private void setUpUpDownButtons(Post post, ImageView upButton, ImageView downButton, TextView nbUps) {
+        upButton.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View v) {
+                if (!post.isUpByUser() && !post.isDownByUser()) {
+                    int newNbUps = post.getNbUps() + 1;
+                    List<String> upScipers = post.getUpScipers();
+                    upScipers.add(post.getUserReading().getSciper());
+
+                    DocumentReference documentReference = db.collection("channels/channel" + post.getChannel().getId() + "/posts").document(post.getId());
+                    documentReference.update(
+                            "nbUps", newNbUps,
+                            "upScipers", upScipers
+                    );
+                    post.setUpByUser(true);
+                    updateUps(post, upButton, downButton);
+                    nbUps.setText("" + (Integer.parseInt(nbUps.getText().toString()) + 1));
+                }
+            }
+        });
+
+        downButton.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View v) {
+                if (!post.isUpByUser() && !post.isDownByUser()) {
+                    int newNbUps = post.getNbUps() - 1;
+                    List<String> downScipers = post.getDownScipers();
+                    downScipers.add(post.getUserReading().getSciper());
+
+                    DocumentReference documentReference = db.collection("channels/channel" + post.getChannel().getId() + "/posts").document(post.getId());
+                    documentReference.update(
+                            "nbUps", newNbUps,
+                            "downScipers", downScipers
+                    );
+                    post.setDownByUser(true);
+                    updateUps(post, upButton, downButton);
+                    nbUps.setText("" + (Integer.parseInt(nbUps.getText().toString()) - 1));
+                }
+            }
+        });
+    }
+
+    private void updateUps(Post post, ImageView upButton, ImageView downButton) {
+        if (post.isUpByUser()) {
             upButton.setImageResource(R.drawable.up_gray);
             downButton.setImageResource(R.drawable.down_transparent);
         }
-        else if (currentPost.isDownByUser()) {
+        else if (post.isDownByUser()) {
             downButton.setImageResource(R.drawable.down_gray);
             upButton.setImageResource(R.drawable.up_transparent);
         }
